requirements.txt
# Deprecated for primary use.
# Prefer `uv sync` with pyproject.toml for the Python 3.10 + gymnasium + mujoco setup.

numpy<2
scipy
torch
fire
termcolor
wandb>=0.22.3
gymnasium[mujoco]>=1.0.0
h5py
